requirements.txt
Django==3.2.10
djangorestframework==3.12.4
drf-spectacular==0.21.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
